+++ CandiesSellerWPF/CandiesSellerWCF/CandiesSellerWCF/Data/stock.cs
@@ -15,11 +15,11 @@ namespace CandiesSellerWCF.Data
     public partial class stock
     {
         public int Id { get; set; }
-        public int Skittles { get; set; }
-        public int cottoncandy { get; set; }
-        public int creamegg { get; set; }
-        public int jawbreaker { get; set; }
-        public int redvinelicornes { get; set; }
-        public int sugardaddy { get; set; }
+        public double Skittles { get; set; }
+        public double cottoncandy { get; set; }
+        public double creamegg { get; set; }
+        public double jawbreaker { get; set; }
+        public double redvinelicornes { get; set; }
+        public double sugardaddy { get; set; }
     }
 }
